cleanup: move experimental `Options` to internal namespace
This class should not be public until its interface is a bit more
finalized, because we don't want a release to go out, implicitly marking
this GA, and making any API tweaks breaking changes.
