+++ google/cloud/google_cloud_cpp_common.bzl
@@ -46,6 +46,7 @@ google_cloud_cpp_common_hdrs = [
     "internal/getenv.h",
     "internal/invoke_result.h",
     "internal/ios_flags_saver.h",
+    "internal/options.h",
     "internal/pagination_range.h",
     "internal/parse_rfc3339.h",
     "internal/port_platform.h",
@@ -61,7 +62,6 @@ google_cloud_cpp_common_hdrs = [
     "kms_key_name.h",
     "log.h",
     "optional.h",
-    "options.h",
     "polling_policy.h",
     "status.h",
     "status_or.h",
