+++ google/cloud/google_cloud_cpp_common_unit_tests.bzl
@@ -31,6 +31,7 @@ google_cloud_cpp_common_unit_tests = [
     "internal/format_time_point_test.cc",
     "internal/future_impl_test.cc",
     "internal/invoke_result_test.cc",
+    "internal/options_test.cc",
     "internal/pagination_range_test.cc",
     "internal/parse_rfc3339_test.cc",
     "internal/random_test.cc",
@@ -42,7 +43,6 @@ google_cloud_cpp_common_unit_tests = [
     "internal/utility_test.cc",
     "kms_key_name_test.cc",
     "log_test.cc",
-    "options_test.cc",
     "status_or_test.cc",
     "status_test.cc",
     "stream_range_test.cc",
